pyproject.toml
[tool.poetry]
name = "com_proj"
version = "0.1.0"
description = ""
authors = ["Lex Chen <clang2you@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.4"
pymysql = "^0.9.3"
pyzmq = "^19.0.0"
colorama = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
